Consolidate tokenizer code in a single module
