+++ .changeset/green-bees-learn.md
@@ -1,0 +1,5 @@
+---
+'@keystonejs/mongo-join-builder': patch
+---
+
+Refactored the package internals for better DX.
